Move entity name rendering into more general EntityRenderers module
