+++ src/Entity.h
@@ -71,7 +71,7 @@ struct EntityVTABLE {
 	void (*SetLocation)(struct Entity* e, struct LocationUpdate* update);
 	PackedCol (*GetCol)(struct Entity* e);
 	void (*RenderModel)(struct Entity* e, double deltaTime, float t);
-	void (*RenderName)(struct Entity* e);
+	cc_bool (*ShouldRenderName)(struct Entity* e);
 };
 
 /* Skin is still being downloaded asynchronously */
@@ -155,10 +155,6 @@ CC_VAR extern struct _EntitiesData {
 void Entities_Tick(struct ScheduledTask* task);
 /* Renders all entities */
 void Entities_RenderModels(double delta, float t);
-/* Renders the name tags of entities, depending on Entities.NamesMode */
-void Entities_RenderNames(void);
-/* Renders hovered entity name tags (these appears through blocks) */
-void Entities_RenderHoveredNames(void);
 /* Removes the given entity, raising EntityEvents.Removed event */
 void Entities_Remove(EntityID id);
 /* Gets the ID of the closest entity to the given entity */
