+++ src/Program.c
@@ -77,8 +77,8 @@ static int RunProgram(int argc, char** argv) {
 #ifdef _MSC_VER
 	/* NOTE: Make sure to comment this out before pushing a commit */
 	//cc_string rawArgs = String_FromConst("UnknownShadow200 fffff 127.0.0.1 25565");
-	cc_string rawArgs = String_FromConst("UnknownShadow200"); 
-	argsCount = String_UNSAFE_Split(&rawArgs, ' ', args, 4);
+	//cc_string rawArgs = String_FromConst("UnknownShadow200"); 
+	//argsCount = String_UNSAFE_Split(&rawArgs, ' ', args, 4);
 #endif
 
 	if (argsCount == 0) {
